+++ sqlx-core/src/mssql/done.rs
@@ -1,5 +1,3 @@
-use crate::done::Done;
-use crate::mssql::Mssql;
 use std::iter::{Extend, IntoIterator};
 
 #[derive(Debug, Default)]
@@ -7,10 +5,8 @@ pub struct MssqlDone {
     pub(super) rows_affected: u64,
 }
 
-impl Done for MssqlDone {
-    type Database = Mssql;
-
-    fn rows_affected(&self) -> u64 {
+impl MssqlDone {
+    pub fn rows_affected(&self) -> u64 {
         self.rows_affected
     }
 }