+++ sqlx-core/src/mysql/done.rs
@@ -1,5 +1,3 @@
-use crate::done::Done;
-use crate::mysql::MySql;
 use std::iter::{Extend, IntoIterator};
 
 #[derive(Debug, Default)]
@@ -12,12 +10,8 @@ impl MySqlDone {
     pub fn last_insert_id(&self) -> u64 {
         self.last_insert_id
     }
-}
 
-impl Done for MySqlDone {
-    type Database = MySql;
-
-    fn rows_affected(&self) -> u64 {
+    pub fn rows_affected(&self) -> u64 {
         self.rows_affected
     }
 }